# docs/swagger.yaml
definitions:
  handlers.AssignmentInput:
    properties:
      date:
        type: string
      name:
        type: string
      subject_name:
        type: string
      weight:
        type: integer
    type: object
  handlers.AttendanceInput:
    properties:
      student_id:
        type: integer
      subject_name:
        type: string
      visit_day:
        type: string
      visited:
        type: boolean
    type: object
  handlers.GradeInput:
    properties:
      assignment_id:
        type: integer
      mark:
        type: integer
      student_id:
        type: integer
    type: object
  models.Assignment:
    properties:
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      subject_name:
        type: string
      weight:
        type: integer
    type: object
  models.Attendance:
    properties:
      id:
        type: integer
      student_id:
        type: integer
      subject_name:
        type: string
      visit_day:
        type: string
      visited:
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.StudentGPA:
    properties:
      gpa:
        type: number
      student_id:
        type: integer
    type: object
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /assignments:
    get:
      consumes:
      - application/json
      description: Get a list of assignments, optionally filtered by subject
      parameters:
      - description: Filter by Subject Name
        in: query
        name: subject_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Assignment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get assignments
      tags:
      - Assignments
    post:
      consumes:
      - application/json
      description: Add a new assignment for a subject
      parameters:
      - description: Assignment Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AssignmentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create assignment
      tags:
      - Assignments
  /attendance:
    get:
      consumes:
      - application/json
      description: Get attendance records filtered by student ID or subject name
      parameters:
      - description: Filter by Student ID
        in: query
        name: student_id
        type: integer
      - description: Filter by Subject Name
        in: query
        name: subject_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Attendance'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get attendance
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      description: Create a new attendance record for a student
      parameters:
      - description: Attendance Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AttendanceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record attendance
      tags:
      - Attendance
  /attendance/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an attendance record from the database
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete attendance
      tags:
      - Attendance
    patch:
      consumes:
      - application/json
      description: Update an existing attendance record
      parameters:
      - description: Attendance ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Attendance Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AttendanceInput'
      produces:
      - application/json
      responses:
        "200":
          description: Returns status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update attendance
      tags:
      - Attendance
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate using email and password to receive a JWT token
      parameters:
      - description: User Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Returns JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User Registration Details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /grades:
    post:
      consumes:
      - application/json
      description: Record a grade for a specific assignment
      parameters:
      - description: Grade Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.GradeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns created ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create grade
      tags:
      - Grades
...
  /rankings:
    get:
      consumes:
      - application/json
      description: Get student rankings by GPA, filtered by group or subject
      parameters:
      - description: Filter by Group ID
        in: query
        name: group_id
        type: integer
      - description: Filter by Subject Name
        in: query
        name: subject_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentGPA'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get rankings
      tags:
      - Grades
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Get the profile details of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Current User
      tags:
      - Auth
swagger: "2.0"